+++ fbw-a380x/src/systems/instruments/src/SDv2/Pages/Video/VideoPage.tsx
@@ -1,0 +1,36 @@
+import { FSComponent, VNode } from '@microsoft/msfs-sdk';
+import { DestroyableComponent } from 'instruments/src/MsfsAvionicsCommon/DestroyableComponent';
+
+import { PageTitle } from '../Generic/PageTitle';
+
+import '../../../index.scss';
+import { SdPageProps } from '../../SD';
+
+export class VideoPage extends DestroyableComponent<SdPageProps> {
+  private topSvgStyle = this.props.visible.map((v) => `visibility: ${v ? 'visible' : 'hidden'}`);
+
+  onAfterRender(node: VNode): void {
+    super.onAfterRender(node);
+
+    this.subscriptions.push(this.topSvgStyle);
+  }
+
+  render() {
+    return (
+      <svg
+        version="1.1"
+        xmlns="http://www.w3.org/2000/svg"
+        xmlnsXlink="http://www.w3.org/1999/xlink"
+        viewBox="0 0 768 1024"
+        style={this.topSvgStyle}
+      >
+        <PageTitle x={6} y={29}>
+          VIDEO
+        </PageTitle>
+        <text x={384} y={343} className="F26 Amber MiddleAlign">
+          NOT AVAIL
+        </text>
+      </svg>
+    );
+  }
+}