+++ .github/CHANGELOG.md
@@ -81,6 +81,7 @@
 1. [ATSU] Improved Hoppie ACARS connection protocol - @heclak (Heclak)
 1. [ATSU] Fixed a case where "no active atc" is returned but station is online - @heclak (Heclak)
 1. [A32NX/FMS] Fixed TOO STEEP PATH AHEAD message appearing greater than 150nm from destination - @Jonny23787 (Jonathan)
+1. [A380X/SD] Port VIDEO SD page to SDv2 - @Jonny23787 (Jonathan)
 
 ## 0.13.0